New magisk mode added that can be triggered from the about tab and pressing the create magisk module button. This will generate the magisk module, uninstall the app and reboot your phone. The magisk module should allow all services to run without needing the notification as android hopefully wont kill a system app.
To uninstall simply remove the magisk module from magisk manager.

NOTE: This is completely optional the functionality is identical for the magisk version as the default.

NOTE: This is a testing feature so there could be some issues. I will test it for about a week before uploading to the playstore if everything is ok